+++ src/core/hook/hook-command-run.h
@@ -34,6 +34,10 @@ struct t_hook_command_run
 {
     t_hook_callback_command_run *callback; /* command_run callback          */
     char *command;                     /* name of command (without '/')     */
+
+    int keep_spaces_right;             /* if set to 1: don't strip trailing */
+                                       /* spaces in args when the command   */
+                                       /* is executed                       */
 };
 
 extern char *hook_command_run_get_description (struct t_hook *hook);