+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - core: add option `-v` to display upgrades in command `/version`
+- api: add property `keep_spaces_right` in function hook_set to keep trailing spaces in command arguments
 
 ### Added